[FEATURE] Implemented a sending queue
Instead of using a serialized array of UIDs the sending of emails
to recipients gets now accomplished via a send queue table. This
feature starts to introduce domain models and repositories for
for many of the tables used by direct mail.

This commit is currently untestet and will most probably not run.
I wrote the whole stuff during the last two weeks without doing a
single test run.